util folder refactor
